analyzer: move eth into util/eth
